Log: DoubleClick log item to apply user-defined offset and scalar
When plotting data it is sometimes nice to add a scalar and/or an offset. Good examples are when comparing GPS alt vs baro alt vs lidar alt.
Any DataFlash .bin/.log logged value can have an arbitrary scalar and/or offset applied when plotted.
Enter modifier then value, they are applied in the order you provide. Modifiers are x + - /
Example: Convert cm to to m with an offset of 50: '/100 +50' or 'x0.01 +50' or '*0.01,+50'

Log: Optimize constructor for faster graphing