Added CameraJoyStick
Added a CameraJoystick class so enable a second joystick to be added to
control camera operations (MPO).  This MPO control has safety features
that are configurable to only allow it to override the main controller
(AVO) if it hits a certain threshold.  If that threshold is set to 0,
then the MPO has 100% control and the AVO has 0% over that channel.