Added sysid on each marker when >1 vehicle connected
When only one vehicle is connected there are no visible changes. I've introduced a small check in the "draw all mavs" loop. If there is more than one vehicle connected it displays the sysid below the altitude. At the moment it's very difficult to determine which vehicles are which and this makes it much clearer.

Apologies if I've done something horribly wrong as I have very little idea about Mission Planners internal workings. (or C# for that matter)