MAVLinkInterface: Do not duplicate DigicamControl messages
setDigicamControl method sends 2 equivalent
messages for the purpose of triggering a camera.
For Autopilot modules that support both messages
this will result in 2 trigger requests, which will
likely not be successfull as they are too close together
in time. Regardless of the success of the trigger on
the camera, the 2 triggers will likely be logged
on the autopilot modules leading to difficulties
when geo-referencing logs.

One example of this functionality leading to problems
is in Ardupilot. See [1].

Also the DIGICAM_CONTROL message seems to be deprecated
in recent Ardupilot versions in favor of the command,
which is the version that we now leave.

[1] https://discuss.ardupilot.org/t/duplicate-coordinate-entry-on-ever-single-trigger-cam-in-dataflash-log/17623/9