Extra currentstate variables for plugin use
Currently it is very hard (if not impossible) to add variables to current state from a plugin (cannot display anything in QuickView tab). This PR adds five cs variables (extra1..5) and the possibility to add description to them from setting. (<extra1_desc> etc...) Since these variables are not binding to datasource they can be directly manipulated from a plugin.