Fix application local data path in *nix based OS
In Linux/Unix like operating systems 'MyDocuments' dir does not exists.
However, the mono framework for compatibility purposes points it to the
root user directory instead. That is incorrect from the point of view of
the Linux/Unit standards (see [1] for more detail). The Local Application
Data directory should be used insted. Fix that for operating systems where
platform ID is Unix.

 * [1] https://www.freedesktop.org/wiki/Specifications/basedir-spec/