ConfigRawParams: fix typecast exception
When editing a float param, the cell gets set to a double, instead of a
string, which throws a typecase exception if the cell is edited and
then "None Default" is clicked. I fix this on both ends, setting the
cell to a string, and using ToString everywhere instead of typecasting.