ADSB: fix deadlock issue
The call to MainV2.comPort.sendPacket waits for MAVLinkInterface.objlock
to release. This can cause a deadlock if the MAVLinkInterface thread
has also called adsb_UpdatePlanePosition, and is waiting on this thread
to release adsblock. adsblock is only needed for accessing the
adsbPlanes list, so this commit restricts the lock to only where it is
needed.