HUD/PrearmStatus: fix failure messages
When prearm checks are failing, the HUD displays the last failure message,
even if that message is very stale. This changes it so it displays the
latest failure message since the last time it passed the prearm check. The
prearm failure messages were also moved to the lowest priority, so that they
would not mask other sensor failures.

The PrearmStatus window had a similar problem, where it would show all
failure messages briefly when it first launched, instead of the latest
relevant ones.

Also, a drive-by fix of where the messageHigh setter function sets the
_messageHighTime variable, which was causing flickers.