GMapMarkerBase: add options for transparent MAVs
This allows the user to select whether other MAVs, or other instances of
the same MAV on redundant links, should be hidden from the map, or
displayed as transparent to be less distracting, or displayed as normal.
The setting defaults to normal.

Additionally, added settings on the Config > Planner page to allow
control over settings that were previously only available with config
hacking.