gps base: add automatic base station configuration for Septentrio (#3324)
* rtk/gps: add automatic base setup for Septentrio receivers

Make the automatic base setup more generic so it can also be used with
other receivers. Add configuration options for Septentrio receivers.

* rtk/gps: use `beginInvoke()` inside async functions

Use `beginInvoke()` inside asynchronous functions to update the UI from
the UI thread.

* rtk/gps: fix quick double click on "connect"

When the user quickly clicked the "connect" button twice, the task to
connect to the receiver would be executed twice, causing problems.
Disable the button until the connection process is finished.