ConnectionOptions fixes (#3254)
* ConnectionOptions: wait to fetch params until added

The SerialReader thread does nothing to our new port until it is added
to the list of Comports, but we cannot add it to that list until the
BaseStream is open, or the SerialReader will dispose of the port.

* MAVLinkInterface: apply theme to progress form

The theme would show correctly for the main connection button, but if
anything else, like ConnectionOptions, launched the connection, the
progress bar would show up unthemed.