Quaternion.cs: overhaul and document
This library was originally copy-pasted from ArduPilot and ported to C#,
but many mistakes were made in the process. For example, many of the
methods in the C++ implementation take a reference variable as input,
and the C# code was taking a regular input and outputting nothing.

This commit overhauls the whole library, making it more C#-like and
documents all methods.